feat: add semantic interrupt support with audio feedback
- Add comprehensive semantic interrupt keywords for Chinese and English
- Implement interrupt tone playback with three-note sound effect
- Configure ASR context with hotword detection for better interrupts
- Refactor type definitions from config.ts to types.ts for better organization
- Update interrupt mode to 0 and ANS mode to 3 for semantic detection
- Remove avatar configuration (disabled)
- Reduce photo capture timeout from 3s to 1.5s for faster UX
