+++ web/src/utils/photo-capture.ts
@@ -98,7 +98,7 @@ export async function capturePhotoFromVideo(
       result.downloadUrl = downloadUrl
 
       // Show success notification
-      toast.success('📸 Photo captured and uploaded successfully! Camera will close in 3 seconds...')
+      toast.success('📸 Photo captured and uploaded successfully! Camera will close in 1.5 seconds...')
     } catch (error) {
       console.warn('Photo upload failed:', error)
       // Show error notification
